+++ src/Anthropic/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 12.8.0 (2026-02-19)
+
+Full Changelog: [Anthropic-v12.7.0...Anthropic-v12.8.0](https://github.com/anthropics/anthropic-sdk-csharp/compare/Anthropic-v12.7.0...Anthropic-v12.8.0)
+
+### Features
+
+* **api:** Add top-level cache control (automatic caching) ([c294cb3](https://github.com/anthropics/anthropic-sdk-csharp/commit/c294cb3eec22f5ad865261d0a5acaee9e81d635c))
+* **api:** Deprecate haiku-3 ([e087138](https://github.com/anthropics/anthropic-sdk-csharp/commit/e0871384ccc36aa31a0036e8cedbd2344b52522e))
+
 ## 12.7.0 (2026-02-18)
 
 Full Changelog: [Anthropic-v12.6.0...Anthropic-v12.7.0](https://github.com/anthropics/anthropic-sdk-csharp/compare/Anthropic-v12.6.0...Anthropic-v12.7.0)